+++ CHANGELOG.md
@@ -1,13 +1,22 @@
-##  1.0.2
+## 1.0.3
+
+* iterable changed to iInterceptors
+* you can force to support http connection
+
+## 1.0.2
+
 * Solve bugs not get error message from api
 
 ## 1.0.1
+
 * Change Readme
 
 ## 1.0.0
+
 * Use sdk 2.12.0 to above
 * now you can send custom Interceptor
 * remove init function use
+
 ```dart
 // default option
 DioManager.init();
@@ -16,9 +25,11 @@ DioManger.initWithCustomOption();
 ```
 
 ## 0.0.5
+
 * Use sdk 2.19.4 to above
 
 ## 0.0.3
+
 * Solve readme
 
 ## 0.0.2